requirements.txt
django==4.1.3
django-autoslug==1.9.8
django-dotenv
django-htmx
Pillow==9.3.0
